Add Tcl evaluation/escape tests
