+++ requirements_dev.txt
@@ -26,6 +26,7 @@ types-Deprecated
 
 # test
 pytest
+pyfakefs>=5.2.3,<6
 
 # ide
 pylance
